requirements.txt
aiogram>=3.0
requests>=2.0
flask>=2.0
sqlalchemy>=2.0
flask-sqlalchemy>=3.1.1
geopy>=2.4.1
flasksqlalchemybasemodel>=1.2.0
pandas>=2.2.3
openpyxl>=3.1.5
python-dotenv>=1.1.0
alembic>=1.15.2
Flask-Migrate>=4.1.0